backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "house-allotment-api"
version = "0.1.0"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "SQLAlchemy>=2.0.32",
  "alembic>=1.13.2",
  "psycopg2-binary>=2.9.9",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.9",
  "email-validator>=2.1.0.post1"
]
